examples: boris: minor bugfixes and header/impl split for sweeper
