+++ examples/boris/boris_sdc.cpp
@@ -11,6 +11,7 @@
 #include "bindings/wrapper_simple_physics_solver.hpp"
 #include "boris_sweeper.hpp"
 
+
 namespace pfasst
 {
   namespace examples
